+++ crates/feagi-trainer/src/binding/class_decoder.rs
@@ -27,7 +27,7 @@ use crate::contracts::prediction_record::TypedPrediction;
 use crate::error::TrainerError;
 
 /// Motor cortical unit index this decoder reads from for the IRIS slice.
-const IRIS_MOTOR_UNIT: u8 = 0;
+const IRIS_MOTOR_UNIT: u16 = 0;
 
 /// Stateless selector that decodes a class index from per-class channel activations.
 #[derive(Debug, Clone, Copy, Default)]
